// app/simple-calculator/types.ts
export enum CalculatorKey {
  Number0 = "0",
  Number1 = "1",
  Number2 = "2",
  Number3 = "3",
  Number4 = "4",
  Number5 = "5",
  Number6 = "6",
  Number7 = "7",
  Number8 = "8",
  Number9 = "9",
  Dot = ".",
  Equal = "=",
  Clear = "C",
  OpenBracket = "(",
  CloseBracket = ")",
  Plus = "+",
  Minus = "-",
  Multiply = "x",
  Divide = "/",
  Remain = "%",
}